backend/requirements.txt
# LVS Portal Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pyotp==2.9.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic[email]==2.5.3
qrcode[pil]==7.4.2
aiosqlite==0.19.0
requests==2.31.0
google-cloud-storage==2.14.0
libsql_experimental==0.0.47
